Render a seperate NavigationMenu component
Note: This works, but it is strange.  Without useLocation, the selected tab does not update.  With the hook it does.  The matcher funtion is not even needed.  It's JUST the fact that we call useLocation, which forces this component to re-render, which updates the selected tab.  We should revisit this at some point, and we'll log an issue to do so.  But for now, it works.